% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_score.R
\name{auto_score}
\alias{auto_score}
\title{auto_score}
\usage{
auto_score(folder = choose.dir(), min_threshold = 5000, x_range = c(200,
  500), channels = 1:5, my.ladder = liz600, marker = "mark",
  plotting = FALSE)
}
\arguments{
\item{folder}{string path to a folder with only ABI files}

\item{min_threshold}{integer, default: 5000}

\item{x_range}{integer vector; default: 200:500}

\item{channels}{integer vector; 1:5}

\item{my.ladder}{integer vector: default: LIZ600 values}

\item{marker}{string; a namer for the marker}

\item{plotting}{logical; whether to create plots or not; default: FALSE}
}
\value{
data.frame of scoring results for each file or individual. Two columns for each marker (diploid).
}
\description{
auto_score
}
\examples{

library(FragmanUI)

data(liz600)

res <- auto_score()

}
