% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_binary.R
\name{convert_to_binary}
\alias{convert_to_binary}
\title{Convert To Binary}
\usage{
convert_to_binary(scores = NULL)
}
\arguments{
\item{scores}{data.frame of score results}
}
\value{
a data.frame
}
\description{
Convert To Binary
}
\examples{

# library(FragmanUI)
#
# scores <- auto_score()
# score_bin <- convert_to_binary(scores)
}
