% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liz600.R
\docType{data}
\name{liz600}
\alias{liz600}
\title{Ladder sizes for LIZ 600}
\format{An object of format vector.}
\usage{
data(liz600)
}
\description{
Ladder sizes for LIZ 600
}
\examples{
data(liz600)
}
\references{
(\href{Data from https://www.thermofisher.com/order/catalog/product/4408399.}{Thermo Fischer})
}
\keyword{datasets}
