% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fran.R
\name{run_fran}
\alias{run_fran}
\title{run_fran}
\usage{
run_fran(port = 9876)
}
\arguments{
\item{port}{integer, default 9876}
}
\value{
NULL invisible
}
\description{
Runs a shiny app to analyze sequence fragments.
}
\details{
Uses Fragman in the backend.
}
\examples{

if(interactive()) {
   FragmanUI::run_fran()
}

}
\author{
Reinhard Simon
}
