% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liz600.R
\docType{data}
\name{liz600}
\alias{liz600}
\title{Ladder sizes for LIZ 600}
\format{An object of format vector.}
\usage{
data(liz600)
}
\description{
Ladder sizes for LIZ 600
}
\examples{
data(liz600)
}
\references{
Data from (\href{https://www.thermofisher.com/order/catalog/product/4366589?SID=srch-srp-4366589}{Thermo Fischer LIZ 600 ladder})
}
\keyword{datasets}
